backend/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
Django
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
